src/requirements.txt
psycopg2-binary==2.9.9
pandas==2.0.3
openpyxl==3.1.2
Pillow==10.0.0